+++ src/components/Timer.tsx
@@ -1,3 +1,8 @@
 import React from 'react';
+import { Button } from './Button';
 
-export const Timer = (): JSX.Element => <div>Timer</div>;
+export const Timer = (): JSX.Element => 
+  <div>
+    Timer
+    <Button />
+  </div>;